parser moved to serial-port-parser catalog
