+++ packages/pure/src/device/serial-port-parser/parser.ts
